+++ ee/server/service/scep/scep_proxy_test.go
@@ -1,4 +1,4 @@
-package service
+package scep
 
 import (
 	"context"
