+++ ee/server/service/scep/testing_utils.go
@@ -1,4 +1,4 @@
-package service
+package scep
 
 import (
 	"crypto/x509"
