+++ server/service/integration_certificate_authorities_test.go
@@ -14,7 +14,7 @@ import (
 	"sync/atomic"
 	"testing"
 
-	eeservice "github.com/fleetdm/fleet/v4/ee/server/service"
+	"github.com/fleetdm/fleet/v4/ee/server/service/scep"
 	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	apple_mdm "github.com/fleetdm/fleet/v4/server/mdm/apple"
@@ -43,9 +43,9 @@ func (s *integrationMDMTestSuite) TestBatchApplyCertificateAuthorities() {
 
 	// TODO(hca): test free version disallows batch endpoint
 
-	ndesSCEPServer := eeservice.NewTestSCEPServer(t)
-	ndesAdminServer := eeservice.NewTestNDESAdminServer(t, "mscep_admin_password", http.StatusOK)
-	dynamicChallengeServer := eeservice.NewTestDynamicChallengeServer(t)
+	ndesSCEPServer := scep.NewTestSCEPServer(t)
+	ndesAdminServer := scep.NewTestNDESAdminServer(t, "mscep_admin_password", http.StatusOK)
+	dynamicChallengeServer := scep.NewTestDynamicChallengeServer(t)
 
 	pathRegex := regexp.MustCompile(`^/mpki/api/v2/profile/([a-zA-Z0-9_-]+)$`)
 	mockDigiCertServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
